scripts: Introduce mkstandalone.sh
This is a pretty ugly bash script, wrapped around a couple ugly
bash scripts, and it generates an ugly sh script. But, it gets the
job done. What's the job? Take a unit test and generate a standalone
script that can be run on any appropriate system with an appropriate
qemu. This makes distributing single, pre-compiled, unit tests easy,
even through email, which can be useful for getting test results from
a variety of users, or even for deploying the unit test as utility
(if it works that way) in a distribution. "Packaging" works like shar,
but doesn't use shar, as it's better to avoid the dependency. Can be
used to create just one unit test, or all of them ('make standalone').
The standalone test(s) are placed in ./tests.

Signed-off-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>