arm/tcg-test: add some IRQ numerology
This expands the tcg tests to gather more data about the IRQ latency and
actually count delivered IRQs better. Unfortunately there is no bullet
proof way for the sending CPU to know if the IRQ has been sent so we
rely on heuristics. Each IRQ is sent with an incremented ID so there is
a maximum possible spacing before the CPU sends the next IRQ of the same
type to the other CPU.

This seems to work mostly as expected on real HW under KVM but TCGs
ability to process the IRQ load is woefully limited. As a result the
checking of these numbers only occurs when the test is manually run
passing the "check_irq" parameter.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>